requirements.txt
streamlit==1.38.0
databricks-sdk==0.57.0
psycopg[binary,pool]==3.3.2